+++ chapaev/qt/gameview.cpp
@@ -30,7 +30,7 @@ void GameView::SetCheckers(Game* game, int num)
 {
     //float bound = (rectSize - 2*Checker::radius) / 2.0;
     checkersHolder = new qCheckerHolder(game);
-    for(int i = 0; i < checkersHolder->GetCheckers().size(); i++)
+    for(size_t i = 0; i < checkersHolder->GetCheckers().size(); i++)
     {
         //int r = Checker::radius;
         //int x = checkersHolder->GetCheckers()[i]->GetChecker()->GetPosition().x();
